Style serverlist column
