+++ src/main/java/com/github/cypher/gui/root/settings/settings.fxml
@@ -1,10 +1,23 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import java.lang.String?>
+<?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.AnchorPane?>
+<?import org.kordamp.ikonli.javafx.FontIcon?>
 
 <AnchorPane minHeight="300.0" minWidth="400.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.github.cypher.gui.root.settings.SettingsPresenter">
    <children>
       <Label alignment="CENTER" layoutX="13.0" layoutY="30.0" prefHeight="42.0" prefWidth="376.0" styleClass="h1" text="Settings" AnchorPane.leftAnchor="5.0" AnchorPane.rightAnchor="5.0" AnchorPane.topAnchor="10.0" />
+      <Button layoutX="163.0" layoutY="101.0" mnemonicParsing="false" onAction="#logout" prefHeight="60.0" prefWidth="60.0" style="-fx-min-width: 2px;" wrapText="true">
+         <graphic>
+            <FontIcon iconColor="WHITE" iconLiteral="fa-angle-left" iconSize="40" textAlignment="CENTER" />
+         </graphic>
+         <styleClass>
+            <String fx:value="btn-info" />
+            <String fx:value="btn" />
+            <String fx:value="image-btn" />
+         </styleClass>
+      </Button>
    </children>
 </AnchorPane>
