+++ src/main/java/com/github/cypher/main.css
@@ -28,3 +28,17 @@ AnchorPane {
 	-room-item-background: #e1e1e1;
 	-room-item-selection-color: #09cdda;
 }
+
+/*Colors */
+* {
+	-fx-base: #404040;
+	-fx-lighter: #808080;
+	-fx-soft-popup: #008CBA;
+	-fx-popup: #09CDDA;
+
+}
+
+.base-separator {
+	-fx-background-color: -fx-lighter;
+
+}
